requirements.txt
# Core dependencies for BreakHis classification pipeline
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Data acquisition
kagglehub>=0.2.0

# Progress bars and utilities
tqdm>=4.66.0

# Optional but recommended
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.17.0

# For Jupyter notebook support (if needed)
jupyter>=1.0.0
ipykernel>=6.25.0
gdown

# TensorBoard for PyTorch
tensorboard

# Quantum computing
pennylane
pennylane-lightning

# Pretrained models (TIMM - PyTorch Image Models)
timm>=0.9.0
